Remove unused code in TrainerController, fix typing
* Removes unused SubprocessEnvManager import in trainer_controller
* Removes unused `steps` argument to `TrainerController._save_model`
* Consolidates unnecessary branching for curricula in
  `TrainerController.advance`
* Moves `reward_buffer` into `TFPolicy` from `PPOPolicy` and adds
  `BCTrainer` support so that we don't have a broken interface /
  undefined behavior when BCTrainer is used with curricula.
